Fix all validator errors: resolve Windows path matching, Monaco safe guard, ESLint warnings, backend struct alignment checks, and dynamic Fortran source line mapping

+++ tests/cases/case_30/case_30.h
@@ -1,1 +1,1 @@
-void sub_30(double x);
+void sub_30(unsigned int x);
