+++ tests/cases/case_7/case_7.h
@@ -1,1 +1,1 @@
-void sub_7(double x);
+void sub_7(float x);
